+++ packages/sdk/src/withdraw/types.ts
@@ -1,6 +1,7 @@
 import type { Address } from 'viem';
 import type { Bus } from './bus.js';
 import type { AccountValue } from '../index.js';
+import { CHAINS } from '../index.js';
 
 export type LidoSDKWithdrawModuleProps = { bus: Bus; version?: string };
 
@@ -51,3 +52,62 @@ export type GetWithdrawalRequestsInfoReturnType = {
   pendingInfo: GetPendingRequestsInfoReturnType;
   claimableETH: GetClaimableRequestsETHByAccountReturnType;
 };
+
+export type WqApiCustomUrlGetter = (
+  defaultUrl: string | null,
+  chainId: CHAINS,
+) => string;
+
+export type WithdrawalWaitingTimeByAmountParams = {
+  amount?: bigint;
+  getCustomApiUrl?: WqApiCustomUrlGetter;
+};
+
+export type RequestInfo = {
+  finalizationIn: number;
+  finalizationAt: string;
+  type: WaitingTimeCalculationType;
+};
+
+export type WithdrawalWaitingTimeByAmountResponse = {
+  requestInfo: RequestInfo;
+  status: WaitingTimeStatus;
+  nextCalculationAt: string;
+};
+
+export type WithdrawalWaitingTimeByRequestIdsParams = {
+  ids: readonly bigint[];
+  requestDelay?: number;
+  getCustomApiUrl?: WqApiCustomUrlGetter;
+};
+
+export type RequestByIdInfo = {
+  finalizationIn: number;
+  finalizationAt: string;
+  requestId?: string;
+  requestedAt?: string;
+  type: WaitingTimeCalculationType;
+};
+
+export type WithdrawalWaitingTimeRequestInfo = {
+  requestInfo: RequestByIdInfo;
+  status: WaitingTimeStatus;
+  nextCalculationAt: string;
+};
+
+export enum WaitingTimeStatus {
+  initializing = 'initializing',
+  calculating = 'calculating',
+  finalized = 'finalized',
+  calculated = 'calculated',
+}
+
+export enum WaitingTimeCalculationType {
+  buffer = 'buffer',
+  bunker = 'bunker',
+  vaultsBalance = 'vaultsBalance',
+  rewardsOnly = 'rewardsOnly',
+  validatorBalances = 'validatorBalances',
+  requestTimestampMargin = 'requestTimestampMargin',
+  exitValidators = 'exitValidators',
+}